requirements.txt
langchain==0.3.*
langchain-core==0.3.*
langchain-community==0.3.*
langchain-chroma==0.2.*
langchain-groq==0.2.*
langgraph==0.2.*

chromadb>=0.5.0
sentence-transformers
tiktoken

fastapi
uvicorn
python-multipart
tavily-python
pypdf
unstructured
python-docx
lxml
docx2txt
python-dotenv
tenacity

# Cloud Services
supabase
qdrant-client
redis
